Remove unneseccary files
