--- inference_output_manifest_filepath.json
@@ -1,3 +1,0 @@
-{
-    "inference_command": " && python /workspace/nemo/examples/asr/transcribe_speech_parallel.py --config-path /lustre/fsw/portfolios/convai/users/ntadevosyan/results/smth/configs --config-name modified_config_0.yaml predict_ds.shuffle=True && python /workspace/nemo/examples/asr/transcribe_speech_parallel.py --config-path /lustre/fsw/portfolios/convai/users/ntadevosyan/results/smth/configs --config-name modified_config_1.yaml predict_ds.shuffle=True && python /workspace/nemo/scripts/pseudo_labeling/write_transcribed_files.py --prediction_filepaths /lustre/fsw/portfolios/convai/users/ntadevosyan/draco_data/datasets/pt/tarred_unlabeled_youtube_2/bucket1/sharded_manifests /lustre/fsw/portfolios/convai/users/ntadevosyan/draco_data/datasets/pt/tarred_unlabeled_youtube_2/bucket2/sharded_manifests  --is_tarred"
-}
